test: update and rename files
